third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Element-innerHTML-expected.txt

This is a testharness.js-based test.
[FAIL] innerHTML on a disconnected element should use the scoped registry it was created with
  assert_true: expected true got false
[FAIL] nested descendants in innerHTML on a disconnected element should use the scoped registry the element was created with
  assert_true: expected true got false
[FAIL] innerHTML on a disconnected element should use the scoped registry it was created with when parsing a simple HTML
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] innerHTML on an inserted element should continue to use the scoped registry it was created with
  assert_true: expected true got false
Harness: the test ran to completion.
